Add Repeat and Optional types
Since "repeat" is not a concept built into the SRLanguageModel stuff,
we achieve it by composing Paths into a Choice. This means that
you must provide an `atLeast` and `atMost` parameter to the Repeat.
For most cases, you could probably get by with passing some arbitrarily
large parameter if you truly need an "infinite" repeat, but be aware
that we're building arrays, so... probably don't want it to be too big

The "Optional" type is simply a convenience subclass of Repeat
that's `atLeast: 0, atMost: 1`. `atLeast: 0` is implemented by an
empty Path.

Unfortunately, there seems to be some weird interaction with having
a Path containing something after a repeat. For example:

```
var num = SGChoice(pickFromStrings: ["one", "two", "three"])
var repeat = SGRepeat(repeat: num, atMost: 3)
SGPath(path: [repeat, SGWord(from: "four")])
```

You would expect it to recognize "one one four", but it doesn't!
It does, however, recognize "one four." And, if you wrap the "four"
in `SGOptional`, it can recognize "one" by itself and "one four,"
but still not "one one four."